53
5 Feb 2023 18:33:08

Implemented Brackeys's custom Wave Spawner that uses enums and made it work for the game. Still needs work to have multiple enemies spawn, and need to use weights for the spawning. The wave counter UI is currently disabled. The problem of mobs not spawning upon retry is gone for now.