62
8 Feb 2023 21:45:16

MP Towers can now be placed on the proper node that supports it (MPTNode). Mismatching a 'tower type' with a 'node type' will not allow the player to place a tower.

Needs some work to make it seamless.

Work on TowerBlueprint and ShopPage is needed, so that it gets data from Tower or MPTower instead.