63
8 Feb 2023 23:31:09

Removed TowerBlueprint class, and now only relying on the Tower and MPTower scripts. Overhauled the affected scripts that previously used TowerBlueprint.

ShopButtons now get data from the assigned towers. However currently purchasing towers needs to be manually assigned to each button. Fix it so that the button sends the data to the ShopPage script.

When hovering on the nodes, if there is a mismatch in tower type (isMPTower) and node type (forMPTowers), then an error color will show up. Else, it will show a normal hover color.

Upgrading and selling need some rewriting of code to use the values in Tower and MPTower scripts.