Burnout causes towers to have reduced damage.
Damage stat was moved from the projectile to the tower itself. The tower gives the projectile the damage it needs upon instantiating. This allows for the damage of the tower to be modified.

Upon Burnout state, the damage will be reduced to 50%. This can be changed within the PlayerStats script.

Made most stat variables of the Tower script has been made private.