Included notification text in game, and added QOL features to tower buttons.
Shop buttons no longer selectable when player has insufficient TP for the tower.

Upgrade buttons grey out when player has insufficient TP for the upgrade.

Notification text in the level is implemented and relies on using event listeners.

Fixed bug of game being 'frozen' with no way out if the card selection pops up while player is in burnout.

Fixed bug that Tutoray is assigned as a splash tower instead of laser.